Update Dockerfile
We now use python:3.8.1-buster as baseimage. Without the gobject
dependency we can use the latest python version, without running
into trouble to not find the gobject bindings for that version.

This also allows us, to use the latest version of pymupdf, as
the only available wheel for buster default python 3.7.3 was
1.16.6. Now: 1.16.11